Laravel 10, Vue3 and Inertia v1

+++ tests/Feature/DeleteAccountTest.php
@@ -4,14 +4,21 @@ namespace Tests\Feature;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Laravel\Jetstream\Features;
 use Tests\TestCase;
 
 class DeleteAccountTest extends TestCase
 {
     use RefreshDatabase;
 
-    public function test_user_accounts_can_be_deleted()
+    public function test_user_accounts_can_be_deleted(): void
     {
+        if (! Features::hasAccountDeletionFeatures()) {
+            $this->markTestSkipped('Account deletion is not enabled.');
+
+            return;
+        }
+
         $this->actingAs($user = User::factory()->create());
 
         $response = $this->delete('/user', [
@@ -21,8 +28,14 @@ class DeleteAccountTest extends TestCase
         $this->assertNull($user->fresh());
     }
 
-    public function test_correct_password_must_be_provided_before_account_can_be_deleted()
+    public function test_correct_password_must_be_provided_before_account_can_be_deleted(): void
     {
+        if (! Features::hasAccountDeletionFeatures()) {
+            $this->markTestSkipped('Account deletion is not enabled.');
+
+            return;
+        }
+
         $this->actingAs($user = User::factory()->create());
 
         $response = $this->delete('/user', [
